HB2-65 move custom DB fields to fields.py
